requirements.txt
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
tqdm>=4.66.0

torch

transformers>=4.40.0
datasets>=2.20.0
evaluate>=0.4.1
accelerate>=0.30.0
safetensors>=0.4.3

fastapi
uvicorn

tensorflow
tensorflow-hub

Pillow>=10.0.0
opencv-python>=4.9.0.80
soundfile>=0.12.1
librosa>=0.10.2
jiwer>=3.0.0
